DB dao stubs and entities

+++ application/src/main/java/to/charlie/basecamp/domain/model/dto/workout/CreateWorkoutRequest.java
@@ -13,7 +13,7 @@ public record CreateWorkoutRequest(
         String type,
         Instant startDate,
         Instant endDate,
-        long durationSeconds,
+        double durationSeconds,
         Double distanceM,
         Double elevationGainM,
         Double elevationLossM,
